+++ app/components/UI/AddCustomToken/index.js
@@ -7,6 +7,7 @@ import { strings } from '../../../../locales/i18n';
 import { isValidAddress } from 'ethereumjs-util';
 import ActionView from '../ActionView';
 import { isSmartContractAddress } from '../../../util/transactions';
+import AnalyticsV2 from '../../../util/analyticsV2';
 
 const styles = StyleSheet.create({
 	wrapper: {
@@ -50,11 +51,31 @@ export default class AddCustomToken extends PureComponent {
 		navigation: PropTypes.object
 	};
 
+	getAnalyticsParams = () => {
+		try {
+			const { NetworkController } = Engine.context;
+			const { chainId, type } = NetworkController?.state?.provider || {};
+			const { address, symbol } = this.state;
+			return {
+				token_address: address,
+				token_symbol: symbol,
+				network_name: type,
+				chain_id: chainId,
+				source: 'Custom token'
+			};
+		} catch (error) {
+			return {};
+		}
+	};
+
 	addToken = async () => {
 		if (!(await this.validateCustomToken())) return;
 		const { AssetsController } = Engine.context;
 		const { address, symbol, decimals } = this.state;
 		await AssetsController.addToken(address, symbol, decimals);
+
+		AnalyticsV2.trackEvent(AnalyticsV2.ANALYTICS_EVENTS.TOKEN_ADDED, this.getAnalyticsParams());
+
 		// Clear state before closing
 		this.setState(
 			{
@@ -104,7 +125,9 @@ export default class AddCustomToken extends PureComponent {
 		let validated = true;
 		const address = this.state.address;
 		const isValidTokenAddress = isValidAddress(address);
-		const toSmartContract = isValidTokenAddress && (await isSmartContractAddress(address));
+		const { NetworkController } = Engine.context;
+		const { chainId } = NetworkController?.state?.provider || {};
+		const toSmartContract = isValidTokenAddress && (await isSmartContractAddress(address, chainId));
 		if (address.length === 0) {
 			this.setState({ warningAddress: strings('token.address_cant_be_empty') });
 			validated = false;