+++ app/components/Views/ResetPassword/index.js
@@ -27,7 +27,7 @@ import Engine from '../../../core/Engine';
 import Device from '../../../util/Device';
 import { colors, fontStyles, baseStyles } from '../../../styles/common';
 import { strings } from '../../../../locales/i18n';
-import { getOnboardingNavbarOptions } from '../../UI/Navbar';
+import { getNavigationOptionsTitle } from '../../UI/Navbar';
 import SecureKeychain from '../../../core/SecureKeychain';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import AppConstants from '../../../core/AppConstants';
@@ -37,6 +37,7 @@ import { ONBOARDING, PREVIOUS_SCREEN } from '../../../constants/navigation';
 import { EXISTING_USER, TRUE, BIOMETRY_CHOICE_DISABLED } from '../../../constants/storage';
 import { getPasswordStrengthWord, passwordRequirementsMet } from '../../../util/password';
 import NotificationManager from '../../../core/NotificationManager';
+import { syncPrefs } from '../../../util/sync';
 
 const styles = StyleSheet.create({
 	mainWrapper: {
@@ -251,7 +252,8 @@ const CONFIRM_PASSWORD = 'confirm_password';
  * View where users can set their password for the first time
  */
 class ResetPassword extends PureComponent {
-	static navigationOptions = ({ navigation }) => getOnboardingNavbarOptions(navigation);
+	static navigationOptions = ({ navigation }) =>
+		getNavigationOptionsTitle(strings('password_reset.change_password'), navigation);
 
 	static propTypes = {
 		/**
@@ -400,6 +402,7 @@ class ResetPassword extends PureComponent {
 		const { originalPassword, password: newPassword } = this.state;
 		const { KeyringController, PreferencesController } = Engine.context;
 		const seedPhrase = await this.getSeedPhrase();
+		const oldPrefs = PreferencesController.state;
 
 		let importedAccounts = [];
 		try {
@@ -426,7 +429,6 @@ class ResetPassword extends PureComponent {
 		const hdKeyring = KeyringController.state.keyrings[0];
 		const existingAccountCount = hdKeyring.accounts.length;
 		const selectedAddress = this.props.selectedAddress;
-		let preferencesControllerState = PreferencesController.state;
 
 		// Create previous accounts again
 		for (let i = 0; i < existingAccountCount - 1; i++) {
@@ -442,16 +444,17 @@ class ResetPassword extends PureComponent {
 			Logger.error(e, 'error while trying to import accounts on recreate vault');
 		}
 
-		// Reset preferencesControllerState
-		preferencesControllerState = PreferencesController.state;
+		//Persist old account/identities names
+		const preferencesControllerState = PreferencesController.state;
+		const prefUpdates = syncPrefs(oldPrefs, preferencesControllerState);
 
 		// Set preferencesControllerState again
-		await PreferencesController.update(preferencesControllerState);
+		await PreferencesController.update(prefUpdates);
 		// Reselect previous selected account if still available
 		if (hdKeyring.accounts.includes(selectedAddress)) {
 			PreferencesController.setSelectedAddress(selectedAddress);
 		} else {
-			PreferencesController.setSelectedAddress(hdKeyring[0]);
+			PreferencesController.setSelectedAddress(hdKeyring.accounts[0]);
 		}
 	};