+++ app/components/Views/Settings/NetworksSettings/NetworkSettings/index.js
@@ -17,6 +17,7 @@ import { jsonRpcRequest } from '../../../../../util/jsonRpcRequest';
 import Logger from '../../../../../util/Logger';
 import { isPrefixedFormattedHexString } from '../../../../../util/number';
 import AppConstants from '../../../../../core/AppConstants';
+import AnalyticsV2 from '../../../../../util/analyticsV2';
 
 const styles = StyleSheet.create({
 	wrapper: {
@@ -268,6 +269,17 @@ class NetworkSettings extends PureComponent {
 				blockExplorerUrl
 			});
 			NetworkController.setRpcTarget(url.href, decimalChainId, ticker, nickname);
+
+			const analyticsParamsAdd = {
+				rpc_url: url.href,
+				chain_id: decimalChainId,
+				source: 'Settings',
+				symbol: ticker,
+				block_explorer_url: blockExplorerUrl,
+				network_name: 'rpc'
+			};
+			AnalyticsV2.trackEvent(AnalyticsV2.ANALYTICS_EVENTS.NETWORK_ADDED, analyticsParamsAdd);
+
 			navigation.navigate('WalletView');
 		}
 	};